backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
captcha
pillow
reportlab
pypdf
google-generativeai
pandas
